Remove old utils file
