+++ utils/formatCurrency.js
@@ -1,0 +1,9 @@
+import { parseNum } from './parseNum.js';
+
+export function formatCurrency(value) {
+    const num = parseNum(value);
+    return num.toLocaleString('es-ES', {
+        minimumFractionDigits: 2,
+        maximumFractionDigits: 2
+    });
+}
