+++ utils/formatDate.js
@@ -1,0 +1,4 @@
+export function formatDate(dateStr) {
+    const date = new Date(dateStr + 'T00:00:00');
+    return date.toLocaleDateString('es-ES');
+}
