+++ utils/getTodayString.js
@@ -1,0 +1,4 @@
+export function getTodayString() {
+    const today = new Date();
+    return today.toISOString().split('T')[0];
+}
